requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.5.4
cryptography==42.0.8
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-bootstrap5==24.2
django-filter==24.2
django-taggit==5.0.1
djangorestframework==3.15.1
drf-spectacular==0.27.2
filelock==3.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
oauthlib==3.2.2
pillow==10.3.0
platformdirs==4.2.1
psycopg==3.1.19
psycopg-binary==3.1.19
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rich-menu==0.1.5
rpds-py==0.18.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.1
